fix script new problems
