+++ src/main/java/net/ccbluex/liquidbounce/value/BlockValue.kt
@@ -1,0 +1,6 @@
+package net.ccbluex.liquidbounce.value
+
+/**
+ * Block value represents a value with a block
+ */
+class BlockValue(name: String, value: Int) : IntegerValue(name, value, 1, 197)
