+++ src/main/java/net/ccbluex/liquidbounce/value/BoolValue.kt
@@ -1,0 +1,16 @@
+package net.ccbluex.liquidbounce.value
+
+import com.google.gson.JsonElement
+import com.google.gson.JsonPrimitive
+
+/**
+ * Bool value represents a value with a boolean
+ */
+open class BoolValue(name: String, value: Boolean) : Value<Boolean>(name, value) {
+    override fun toJson() = JsonPrimitive(value)
+
+    override fun fromJson(element: JsonElement) {
+        if (element.isJsonPrimitive)
+            value = element.asBoolean || element.asString.equals("true", ignoreCase = true)
+    }
+}
