+++ src/main/java/net/ccbluex/liquidbounce/value/FloatValue.kt
@@ -1,0 +1,21 @@
+package net.ccbluex.liquidbounce.value
+
+import com.google.gson.JsonElement
+import com.google.gson.JsonPrimitive
+
+/**
+ * Float value represents a value with a float
+ */
+open class FloatValue(name: String, value: Float, val minimum: Float = 0F, val maximum: Float = Float.MAX_VALUE) : Value<Float>(name, value) {
+
+    fun set(newValue: Number) {
+        set(newValue.toFloat())
+    }
+
+    override fun toJson() = JsonPrimitive(value)
+
+    override fun fromJson(element: JsonElement) {
+        if (element.isJsonPrimitive)
+            value = element.asFloat
+    }
+}
