+++ src/main/java/net/ccbluex/liquidbounce/value/FontValue.kt
@@ -1,0 +1,25 @@
+package net.ccbluex.liquidbounce.value
+
+import com.google.gson.JsonElement
+import com.google.gson.JsonObject
+import net.ccbluex.liquidbounce.ui.font.Fonts
+import net.minecraft.client.gui.FontRenderer
+
+/**
+ * Font value represents a value with a font
+ */
+class FontValue(valueName: String, value: FontRenderer) : Value<FontRenderer>(valueName, value) {
+    override fun toJson(): JsonElement? {
+        val fontDetails = Fonts.getFontDetails(value) ?: return null
+        val valueObject = JsonObject()
+        valueObject.addProperty("fontName", fontDetails[0] as String)
+        valueObject.addProperty("fontSize", fontDetails[1] as Int)
+        return valueObject
+    }
+
+    override fun fromJson(element: JsonElement) {
+        if (!element.isJsonObject) return
+        val valueObject = element.asJsonObject
+        value = Fonts.getFontRenderer(valueObject["fontName"].asString, valueObject["fontSize"].asInt)
+    }
+}
