+++ src/main/java/net/ccbluex/liquidbounce/value/IntegerValue.kt
@@ -1,0 +1,21 @@
+package net.ccbluex.liquidbounce.value
+
+import com.google.gson.JsonElement
+import com.google.gson.JsonPrimitive
+
+/**
+ * Integer value represents a value with a integer
+ */
+open class IntegerValue(name: String, value: Int, val minimum: Int = 0, val maximum: Int = Integer.MAX_VALUE) : Value<Int>(name, value) {
+
+    fun set(newValue: Number) {
+        set(newValue.toInt())
+    }
+
+    override fun toJson() = JsonPrimitive(value)
+
+    override fun fromJson(element: JsonElement) {
+        if (element.isJsonPrimitive)
+            value = element.asInt
+    }
+}
