+++ src/main/java/net/ccbluex/liquidbounce/value/ListValue.kt
@@ -1,0 +1,36 @@
+package net.ccbluex.liquidbounce.value
+
+import com.google.gson.JsonElement
+import com.google.gson.JsonPrimitive
+import java.util.*
+
+/**
+ * List value represents a selectable list of values
+ */
+open class ListValue(name: String, val values: Array<String>, value: String) : Value<String>(name, value) {
+    @JvmField
+    var openList = false
+
+    init {
+        this.value = value
+    }
+
+    operator fun contains(string: String?): Boolean {
+        return Arrays.stream(values).anyMatch { s: String -> s.equals(string, ignoreCase = true) }
+    }
+
+    override fun changeValue(value: String) {
+        for (element in values) {
+            if (element.equals(value, ignoreCase = true)) {
+                this.value = element
+                break
+            }
+        }
+    }
+
+    override fun toJson() = JsonPrimitive(value)
+
+    override fun fromJson(element: JsonElement) {
+        if (element.isJsonPrimitive) changeValue(element.asString)
+    }
+}
