+++ src/main/java/net/ccbluex/liquidbounce/value/TextValue.kt
@@ -1,0 +1,16 @@
+package net.ccbluex.liquidbounce.value
+
+import com.google.gson.JsonElement
+import com.google.gson.JsonPrimitive
+
+/**
+ * Text value represents a value with a string
+ */
+open class TextValue(name: String, value: String) : Value<String>(name, value) {
+    override fun toJson() = JsonPrimitive(value)
+
+    override fun fromJson(element: JsonElement) {
+        if (element.isJsonPrimitive)
+            value = element.asString
+    }
+}
