+++ src/main/java/net/ccbluex/liquidbounce/value/Value.kt
@@ -6,19 +6,10 @@
 package net.ccbluex.liquidbounce.value
 
 import com.google.gson.JsonElement
-import com.google.gson.JsonObject
-import com.google.gson.JsonPrimitive
 import net.ccbluex.liquidbounce.LiquidBounce
-import net.ccbluex.liquidbounce.ui.font.Fonts
 import net.ccbluex.liquidbounce.utils.ClientUtils
-import net.minecraft.client.gui.FontRenderer
-import net.minecraftforge.fml.relauncher.Side
-import net.minecraftforge.fml.relauncher.SideOnly
-import java.util.*
 
-@SideOnly(Side.CLIENT)
 abstract class Value<T>(val name: String, protected var value: T) {
-
     fun set(newValue: T) {
         if (newValue == value) return
 
@@ -45,127 +36,4 @@ abstract class Value<T>(val name: String, protected var value: T) {
 
     protected open fun onChange(oldValue: T, newValue: T) {}
     protected open fun onChanged(oldValue: T, newValue: T) {}
-
-}
-
-/**
- * Bool value represents a value with a boolean
- */
-open class BoolValue(name: String, value: Boolean) : Value<Boolean>(name, value) {
-
-    override fun toJson() = JsonPrimitive(value)
-
-    override fun fromJson(element: JsonElement) {
-        if (element.isJsonPrimitive)
-            value = element.asBoolean || element.asString.equals("true", ignoreCase = true)
-    }
-
-}
-
-/**
- * Integer value represents a value with a integer
- */
-open class IntegerValue(name: String, value: Int, val minimum: Int = 0, val maximum: Int = Integer.MAX_VALUE)
-    : Value<Int>(name, value) {
-
-    fun set(newValue: Number) {
-        set(newValue.toInt())
-    }
-
-    override fun toJson() = JsonPrimitive(value)
-
-    override fun fromJson(element: JsonElement) {
-        if (element.isJsonPrimitive)
-            value = element.asInt
-    }
-
-}
-
-/**
- * Float value represents a value with a float
- */
-open class FloatValue(name: String, value: Float, val minimum: Float = 0F, val maximum: Float = Float.MAX_VALUE)
-    : Value<Float>(name, value) {
-
-    fun set(newValue: Number) {
-        set(newValue.toFloat())
-    }
-
-    override fun toJson() = JsonPrimitive(value)
-
-    override fun fromJson(element: JsonElement) {
-        if (element.isJsonPrimitive)
-            value = element.asFloat
-    }
-
-}
-
-/**
- * Text value represents a value with a string
- */
-open class TextValue(name: String, value: String) : Value<String>(name, value) {
-
-    override fun toJson() = JsonPrimitive(value)
-
-    override fun fromJson(element: JsonElement) {
-        if (element.isJsonPrimitive)
-            value = element.asString
-    }
-}
-
-/**
- * Font value represents a value with a font
- */
-class FontValue(valueName: String, value: FontRenderer) : Value<FontRenderer>(valueName, value) {
-
-    override fun toJson(): JsonElement? {
-        val fontDetails = Fonts.getFontDetails(value) ?: return null
-        val valueObject = JsonObject()
-        valueObject.addProperty("fontName", fontDetails[0] as String)
-        valueObject.addProperty("fontSize", fontDetails[1] as Int)
-        return valueObject
-    }
-
-    override fun fromJson(element: JsonElement) {
-        if (!element.isJsonObject) return
-        val valueObject = element.asJsonObject
-        value = Fonts.getFontRenderer(valueObject["fontName"].asString, valueObject["fontSize"].asInt)
-    }
-}
-
-/**
- * Block value represents a value with a block
- */
-class BlockValue(name: String, value: Int) : IntegerValue(name, value, 1, 197)
-
-/**
- * List value represents a selectable list of values
- */
-open class ListValue(name: String, val values: Array<String>, value: String) : Value<String>(name, value) {
-
-    @JvmField
-    var openList = false
-
-    init {
-        this.value = value
-    }
-
-    operator fun contains(string: String?): Boolean {
-        return Arrays.stream(values).anyMatch { s: String -> s.equals(string, ignoreCase = true) }
-    }
-
-    override fun changeValue(value: String) {
-        for (element in values) {
-            if (element.equals(value, ignoreCase = true)) {
-                this.value = element
-                break
-            }
-        }
-    }
-
-    override fun toJson() = JsonPrimitive(value)
-
-    override fun fromJson(element: JsonElement) {
-        if (element.isJsonPrimitive) changeValue(element.asString)
-    }
 }
